tests/wpt/mozilla/tests/mozilla/element_attributes.html

<html>
<head id="foo">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div></div>
<script>
test(function() {
  var attrs = [];
  for (var i = 'a'.charCodeAt(0); i != 'z'.charCodeAt(0); i++) {
    document.getElementsByTagName('div')[0].setAttribute(String.fromCharCode(i),
                                                         i.toString());
  }
  var attributes = document.getElementsByTagName('div')[0].attributes;
  for (var i = 0; i < attributes.length; i++) {
    assert_equals(attributes[i].name, String.fromCharCode(i + 'a'.charCodeAt(0)));
    assert_equals(attributes[i].value, (i + 'a'.charCodeAt(0)).toString());
  }
});
</script>
</body>
</html>
